+++ changelog.md
@@ -13,3 +13,13 @@
 ## [2.1.1] - 2025-12-16
 - Correction de l'authentification sur le handler SMTP
 - Ajout des options cc, bcc et attachments sur le handler SMTP
+
+## [2.1.2] - 2026-02-11
+- Désactivation du logger root par défaut
+
+## [2.1.3] - 2026-02-11
+- Ajout methode log dans logger
+
+## [2.2.0] - 2026-02-11
+- Amélioration du render yaml/jinja
+- Ajout du handler discord